+++ lib/guard/puma/runner.rb
@@ -6,49 +6,48 @@ module Guard
 
     MAX_WAIT_COUNT = 20
 
-    attr_reader :options, :control_url, :control_token, :cmd_opts
+    attr_reader :options, :control_url, :control_token, :cmd_opts, :pumactl
 
     def initialize(options)
       @control_token = options.delete(:control_token) { |_| ::Puma::Configuration.random_token }
-      @control = "localhost"
       @control_port = (options.delete(:control_port) || '9293')
-      @control_url = "#{@control}:#{@control_port}"
+      @control_url = "localhost:#{@control_port}"
       @quiet = options.delete(:quiet) { true }
+      @pumactl = options.delete(:pumactl) { false }
       @options = options
 
-      puma_options = if options[:config]
-        {
-          '--config' => options[:config],
-          '--control-token' => @control_token,
-          '--control' => "tcp://#{@control_url}",
-          '--environment' => options[:environment]
-        }
+      puma_options = {
+        puma_options_key(:config) => options[:config],
+        puma_options_key(:control_token) => @control_token,
+        puma_options_key(:control_url) => "tcp://#{@control_url}"
+      }
+      if options[:config]
+        puma_options['--config'] = options[:config]
       else
-        {
-          '--port' => options[:port],
-          '--control-token' => @control_token,
-          '--control' => "tcp://#{@control_url}",
-          '--environment' => options[:environment]
-        }
+        puma_options['--port'] = options[:port]
       end
-      [:bind, :threads].each do |opt|
-        puma_options["--#{opt}"] = options[opt] if options[opt]
-      end
+      %i[bind threads environment]
+        .select { |opt| options[opt] }
+        .each do |opt|
+          if pumactl
+            Compat::UI.warning(
+              "`#{opt}` option is not compatible with `pumactl` option"
+            )
+          else
+            puma_options["--#{opt}"] = options[opt]
+          end
+        end
       puma_options = puma_options.to_a.flatten
-      puma_options << '-q' if @quiet
+      puma_options << '--quiet' if @quiet
       @cmd_opts = puma_options.join ' '
     end
 
     def start
-      if in_windows_cmd?
-        Kernel.system windows_start_cmd
-      else
-        Kernel.system nix_start_cmd
-      end
+      Kernel.system build_command('start')
     end
 
     def halt
-      Net::HTTP.get build_uri('halt')
+      run_puma_command!('halt')
       # server may not have been stopped correctly, but we are halting so who cares.
       return true
     end
@@ -68,8 +67,30 @@ module Guard
 
     private
 
+    PUMA_OPTIONS_KEYS_BY_PUMACTL = {
+      true => {
+        config:      '--config-file',
+        control_url: '--control-url'
+      }.freeze,
+      false => {
+        config:      '--config',
+        control_url: '--control'
+      }.freeze
+    }.freeze
+
+    private_constant :PUMA_OPTIONS_KEYS_BY_PUMACTL
+
+    def puma_options_key(key)
+      keys = PUMA_OPTIONS_KEYS_BY_PUMACTL[@pumactl]
+      keys.fetch(key) { |k| "--#{k.to_s.tr('_', '-')}" }
+    end
+
     def run_puma_command!(cmd)
-      Net::HTTP.get build_uri(cmd)
+      if pumactl
+        Kernel.system build_command(cmd)
+      else
+        Net::HTTP.get build_uri(cmd)
+      end
       return true
     rescue Errno::ECONNREFUSED => e
       # server may not have been started correctly.
@@ -80,12 +101,22 @@ module Guard
       URI "http://#{control_url}/#{cmd}?token=#{control_token}"
     end
 
-    def nix_start_cmd
-      %{sh -c 'cd #{Dir.pwd} && puma #{cmd_opts} &'}
+    def build_command(cmd)
+      puma_cmd = "#{pumactl ? 'pumactl' : 'puma'} #{cmd_opts} #{cmd if pumactl}"
+      background = cmd == 'start'
+      if in_windows_cmd?
+        windows_cmd(puma_cmd, background)
+      else
+        nix_cmd(puma_cmd, background)
+      end
     end
 
-    def windows_start_cmd
-      %{cd "#{Dir.pwd}" && start "" /B cmd /C "puma #{cmd_opts}"}
+    def nix_cmd(puma_cmd, background = false)
+      %(sh -c 'cd #{Dir.pwd} && #{puma_cmd} #{'&' if background}')
+    end
+
+    def windows_cmd(puma_cmd, background = false)
+      %(cd "#{Dir.pwd}" && #{'start "" /B' if background} cmd /C "#{puma_cmd}")
     end
 
     def in_windows_cmd?